# live_v04/openapi.json
{"openapi":"3.1.0","info":{"title":"Dharma Fleet Hub","version":"0.4.0"},"paths":{"/events/{channel}":{"get":{"summary":"Sse Events","operationId":"sse_events_events__channel__get","parameters":[{"name":"channel","in":"path","required":true,"schema":{"type":"string","title":"Channel"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/send":{"post":{"summary":"Send Message","operationId":"send_message_api_send_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/roster":{"get":{"summary":"Api Roster","operationId":"api_roster_api_roster_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/agent/{uid}":{"get":{"summary":"Agent Detail","operationId":"agent_detail_api_agent__uid__get","parameters":[{"name":"uid","in":"path","required":true,"schema":{"type":"string","title":"Uid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dispatch":{"get":{"summary":"Dispatch Test","description":"Test Wayfinder dispatch without sending a message.","operationId":"dispatch_test_api_dispatch_get","parameters":[{"name":"task","in":"query","required":true,"schema":{"type":"string","title":"Task"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat":{"get":{"summary":"Api Chat","operationId":"api_chat_api_chat_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/kanban":{"get":{"summary":"Api Kanban","operationId":"api_kanban_api_kanban_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/status":{"get":{"summary":"Fleet Status","operationId":"fleet_status_api_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SendRequest":{"properties":{"from_":{"type":"string","title":"From","default":"operator"},"text":{"type":"string","title":"Text"},"to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To"}},"type":"object","required":["text"],"title":"SendRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}